requirements.txt
discord.py
requests
scholarly
json